Fix dataview race conditions and add hyperdataset method wrappers
